% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv data: fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that names the file the user wants to import}
}
\value{
This function returns a tibble if the file exists, if not it returns an warning
with a message ("file *filename* does not exist")
}
\description{
This is a simple function that reads data from a .csv file into a data.frame using tbl_df.
}
\examples{
\dontrun{fars_read("accident_2015.csv.bz2")}
}
