% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple csv data using fars_read and make_filename}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a number that represents the year or years of data to import.
the user can import 1 year data or create a sequence like n:m (n<=m), the function will
read the files from year n to m}
}
\value{
This function returns a dataset with with the following structure:
one line per accident with Month and Year of accident as columns
if a cvs file of a year on the function parameters does not existe the funtion will
throw an warning with the message "invalid year: \code{current year}"
}
\description{
Iteractive function that enables the user to read multiple .csv files from a range of years
It will read a group of files on the same folder changing the year inside a range
that was defined by the user
}
\examples{
\dontrun{fars_read_years(2013)}
\dontrun{fars_read_years(c(2013:2015))}

}
