% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Get statistics of accidents of multiple year per month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a number that represents the year or years of data to import.
the user can import 1 year data or create a sequence like n:m (n<=m) and the function will
read the files from year n to m}
}
\value{
This function returns a dataset with the count os accidents of
every month (columns) grouped by year (rows)
}
\description{
Iteractive function that enable obtain the statistics from multiple .csv files from a range of years
It will read a group of files, row_bind all the data into one data.frame and them compute
statistics from the final data set.
}
\examples{
\dontrun{fars_summarize_years(2013)}
\dontrun{fars_summarize_years(2013:2015)}
}
