% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Help function to create a file name string}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a number that represents the year of the data to import.
the input does not need to be numeric, the function will parse the parameter to integer}
}
\value{
This function returns a string, with the following structure: "accident_(year).csv.bz2"
the year format in of the type 2001, 2011, this coincides qith the current design
of the .csv used for the analysis
}
\description{
This is a function that creates a string with the needed structure to be
used in fars_read_years function for iterative importing several .csv files
}
\examples{
\dontrun{make_filename(1999)}
\dontrun{make_filename("2013")}

}
